Cognitive-Search/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cosmos==4.4.0
aiohttp
azure-search-documents==11.4.0b6
azure-identity
PdfReader
PageObject
openai[datalib]==0.27.6
tiktoken
config
pypdf